==> Code/01_UoE_researchers.ipynb <==
import requests       # send requests to web server
from lxml import html # parse HTML
import json           # store data as json file
import re             # regular expressions
import os             # access directories
import pandas as pd   # dataframes
from tqdm import tqdm # create progress bar (for i in tqdm(list))
os.chdir('../Data')

# 1) Scrape the data

### The Business School


college_name = "The Business School"

urls = ["https://business-school.exeter.ac.uk/about/people/economics/",
        "http://business-school.exeter.ac.uk/about/people/finance/",
        "https://business-school.exeter.ac.uk/about/people/management/",
        "https://business-school.exeter.ac.uk/about/people/site/",
        "https://business-school.exeter.ac.uk/about/people/sustainable-futures/",
        "https://business-school.exeter.ac.uk/about/people/pgr/"]

departments = ["Economics", "Finance and Accounting", "Management",
               "SITE (Science, Innovation, Technology, and Entrepreneurship)",
               "Sustainable Futures", "Business School PGR"]

name = []
website = []
role = []
department = []
college = []

for i, url in enumerate(urls):
    page = requests.get(url)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(url)
    for j in range(1,500):
        path_name = '//*[@class="profile_list"]/tr['+str(j)+']/td[1]/a/text()'
        if tree.xpath(path_name) != []:
            name.append(tree.xpath(path_name)[0])
            path_url = '//*[@class="profile_list"]/tr['+str(j)+']/td[1]/a/@href'
            try:
                website.append(tree.xpath(path_url)[0])
            except:
                website.append("") 
            path_role = '//*[@class="profile_list"]/tr['+str(j)+']/td[2]/text()'
            try:
                role.append(tree.xpath(path_role)[0])
            except: 
                role.append("")
            college.append(college_name)
            department.append(departments[i])
        
businessschool = pd.DataFrame(list(zip(college, department, name, role, website)),
                              columns=['college', 'department', 'name', 'role', 'url'])
businessschool

businessschool.department.value_counts()

### College of Engineering, Mathematics and Physical Sciences

college_name = "College of Engineering, Mathematics and Physical Sciences"

urls = ["https://emps.exeter.ac.uk/staff/academic/",
        "https://emps.exeter.ac.uk/staff/researchstaff/",
        "https://emps.exeter.ac.uk/staff/pgrstudents/",
        "https://emps.exeter.ac.uk/staff/honorarystaffvisitors/"]

name = []
website = []
role = []
department = []
college = []

for i, url in enumerate(urls):
    page = requests.get(url)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(url)
    departments = tree.xpath('//*[@class="profile_list"]/tbody/tr[1]/th/text()')
    for j, dep in enumerate(departments):
        j+=1
        for k in range(1,500):
            path_name = '//*[@class="profile_list"]['+str(j)+']/tbody/tr['+str(k)+']/td[1]/a/text()'
            if tree.xpath(path_name) != []:
                name.append(tree.xpath(path_name)[0])
                path_url = '//*[@class="profile_list"]['+str(j)+']/tbody/tr['+str(k)+']/td[1]/a/@href'
                try:
                    website.append(tree.xpath(path_url)[0])
                except:
                    website.append("")           
                path_role = '//*[@class="profile_list"]['+str(j)+']/tbody/tr['+str(k)+']/td[2]/text()'
                try:
                    role.append(tree.xpath(path_role)[0])
                except: 
                    role.append("")
                department.append(dep)
                college.append(college_name)


cemps = pd.DataFrame(list(zip(college, department, name, role, website)),
                     columns=['college', 'department', 'name', 'role', 'url'])
cemps

cemps.department.value_counts()

### College of Humanities

college_name = "College of Humanities"

urls = ["https://humanities.exeter.ac.uk/archaeology/staff/",
        "https://humanities.exeter.ac.uk/arthistory/staff/",
        "https://humanities.exeter.ac.uk/classics/staff/    ",
        "https://humanities.exeter.ac.uk/drama/staff/",
        "https://humanities.exeter.ac.uk/english/staff/",
        "https://humanities.exeter.ac.uk/film/staff/",
        "https://humanities.exeter.ac.uk/history/staff/",
        "https://humanities.exeter.ac.uk/liberalarts/ourpeople/staffprofiles/",
        "https://humanities.exeter.ac.uk/modernlanguages/staff/",
        "https://humanities.exeter.ac.uk/theology/staff/"]

departments = ["Archaeology", "Art History & Visual Culture", "Classics and Ancient History",
               "Drama", "English", "Film Studies", "History", "Liberal Arts", "Modern Languages",
               "Theology and Religion"]

name = []
website = []
role = []
department = []
college = []

for i, url in enumerate(urls):
    page = requests.get(url)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(url)
    for j in range(1,10):
        for k in range(1,500):
            path_name = '//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr['+str(k)+']/td[1]/a/text()'
            if tree.xpath(path_name) != []:
                name.append(tree.xpath(path_name)[0])
                path_url = '//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr['+str(k)+']/td[1]/a/@href'
                try:
                    website.append(tree.xpath(path_url)[0])
                except:
                    website.append("") 
                path_role = '//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr['+str(k)+']/td[2]/text()'
                if departments[i] not in ["Art History & Visual Culture", 
                                          "Classics and Ancient History", 
                                          "English",  "History",
                                          "Liberal Arts", "Modern Languages"]:
                    try:
                        role.append(tree.xpath(path_role)[0])
                    except: 
                        role.append("")
                else:
                    role.append("")
                college.append(college_name)
                department.append(departments[i])
        
humanities = pd.DataFrame(list(zip(college, department, name, role, website)),
                              columns=['college', 'department', 'name', 'role', 'url'])
humanities

humanities.department.value_counts()

### College of Life and Environmental Sciences

college_name = "College of Life and Environmental Sciences"

urls = ["https://biosciences.exeter.ac.uk/cec/staff/",
        "http://geography.exeter.ac.uk/exeter/staff/",
        "https://psychology.exeter.ac.uk/staff/academicstaff/"
        "https://psychology.exeter.ac.uk/staff/research/",
        "https://psychology.exeter.ac.uk/staff/pgrstudents/",
        "https://sshs.exeter.ac.uk/staff/academic/",
        "https://sshs.exeter.ac.uk/staff/research/",
        "https://sshs.exeter.ac.uk/staff/researchstudents/"]

departments = ["Biosciences", "Geography", "Psychology", "Psychology", "Psychology",
               "Sport and Health Sciences", "Sport and Health Sciences", "Sport and Health Sciences"]

name = []
website = []
role = []
department = []
college = []

for i, url in enumerate(urls):
    page = requests.get(url)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(url)
    for j in range(1,500):
        path_name = '//*[@class="profile_list"]/tr['+str(j)+']/td[1]/a/text()'
        if tree.xpath(path_name) != []:
            name.append(tree.xpath(path_name)[0])
            path_url = '//*[@class="profile_list"]/tr['+str(j)+']/td[1]/a/@href'
            try:
                website.append(tree.xpath(path_url)[0])
            except:
                website.append("") 
            path_role = '//*[@class="profile_list"]/tr['+str(j)+']/td[2]/text()'
            try:
                role.append(tree.xpath(path_role)[0])
            except: 
                role.append("")
            college.append(college_name)
            department.append(departments[i])
        
lifesciences = pd.DataFrame(list(zip(college, department, name, role, website)),
                              columns=['college', 'department', 'name', 'role', 'url'])
lifesciences

lifesciences.department.value_counts()

### College of Medicine and Health

college_name = "College of Medicine and Health"

urls = ["https://medicine.exeter.ac.uk/people/academic/",
        "https://medicine.exeter.ac.uk/people/researchstudents/"]

departments = ["Medicine and Health",
               "Medicine and Health"]

name = []
website = []
role = []
department = []
college = []

for i, url in enumerate(urls):
    page = requests.get(url)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(url)
    for j in range(1,600):
        path_name = '//*[@class="profile_list"]/tr['+str(j)+']/td[1]/a/text()'
        if tree.xpath(path_name) != []:
            name.append(tree.xpath(path_name)[0])
            path_url = '//*[@class="profile_list"]/tr['+str(j)+']/td[1]/a/@href'
            try:
                website.append(tree.xpath(path_url)[0])
            except:
                website.append("") 
            path_role = '//*[@class="profile_list"]/tr['+str(j)+']/td[2]/text()'
            try:
                role.append(tree.xpath(path_role)[0])
            except: 
                role.append("")
            college.append(college_name)
            department.append(departments[i])
        
medicine = pd.DataFrame(list(zip(college, department, name, role, website)),
                              columns=['college', 'department', 'name', 'role', 'url'])
medicine

medicine.department.value_counts()

### College of Social Sciences and International Studies

college_name = "College of Social Sciences and International Studies"

urls = ["http://socialsciences.exeter.ac.uk/iais/staff/",
        "http://socialsciences.exeter.ac.uk/law/staff/",
        "http://socialsciences.exeter.ac.uk/politics/staff/",
        "http://socialsciences.exeter.ac.uk/sociology/staff/"]

departments = ["Arab and Islamic Studies", "Law", "Politics",
               "Sociology, Philosophy and Anthropology"]

name = []
website = []
role = []
department = []
college = []

for i, url in enumerate(urls):
    page = requests.get(url)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(url)
    for j in range(1,15):
        for k in range(1,500):
            path_name = '//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr['+str(k)+']/td[1]/a/text()'
            if tree.xpath(path_name) != []:
                name.append(tree.xpath(path_name)[0])
                path_url = '//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr['+str(k)+']/td[1]/a/@href'
                try:
                    website.append(tree.xpath(path_url)[0])
                except:
                    website.append("") 
                trigger =  False
                if len(tree.xpath('//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr/th/text()')) != 0:
                    if 'PhD students' in tree.xpath('//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr/th/text()')[0]:
                        role.append("PhD student")
                        trigger = True
                if not trigger:
                    path_role = '//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr['+str(k)+']/td[2]/p/text()'
                    try:                           
                        role.append(tree.xpath(path_role)[0])                  
                    except:
                        role.append("")
                college.append(college_name)
                department.append(departments[i])
            else:
                path_name = '//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr['+str(k)+']/td[1]/text()'
                if tree.xpath(path_name) != []:
                    name.append(tree.xpath(path_name)[0])
                    path_url = '//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr['+str(k)+']/td[1]/@href'
                    try:
                        website.append(tree.xpath(path_url)[0])
                    except:
                        website.append("")
                    trigger =  False
                    if len(tree.xpath('//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr/th/text()')) != 0:
                        if 'PhD students' in tree.xpath('//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr/th/text()')[0]:
                            role.append("PhD student")
                            trigger = True
                    if not trigger:
                        path_role = '//*[@class="row"]/div[2]/table['+str(j)+']/tbody/tr['+str(k)+']/td[2]/p/text()'
                        try:
                            role.append(tree.xpath(path_role)[0])                  
                        except:
                            role.append("")
                    college.append(college_name)
                    department.append(departments[i])
                
                
urls = ["https://socialsciences.exeter.ac.uk/education/staff/all/"]
departments = ["Education"]

for i, url in enumerate(urls):
    page = requests.get(url)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(url)
    for j in range(1,600):
        path_name = '//*[@class="profile_list"]/tr['+str(j)+']/td[1]/a/text()'
        if tree.xpath(path_name) != []:
            name.append(tree.xpath(path_name)[0])
            path_url = '//*[@class="profile_list"]/tr['+str(j)+']/td[1]/a/@href'
            try:
                website.append(tree.xpath(path_url)[0])
            except:
                website.append("") 
            path_role = '//*[@class="profile_list"]/tr['+str(j)+']/td[2]/text()'
            try:
                role.append(tree.xpath(path_role)[0])
            except: 
                role.append("")
            college.append(college_name)
            department.append(departments[i])
            
urls = ["http://www.exeter.ac.uk/strategy/people/"]
departments = ["Strategy and Security"]

for i, url in enumerate(urls):
    page = requests.get(url)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(url)
    for j in range(1,6):
        for k in range(1,500):
            path_name = '//*[@class="panel-group accordion"]/div['+str(j)+']/div[2]/div/table/tbody/tr['+str(k)+']/td[1]/a/text()'
            if tree.xpath(path_name) != []:
                name.append(tree.xpath(path_name)[0])
                path_url = '//*[@class="panel-group accordion"]/div['+str(j)+']/div[2]/div/table/tbody/tr['+str(k)+']/td[1]/a/@href'
                try:
                    website.append(tree.xpath(path_url)[0])
                except:
                    website.append("") 
                path_role = '//*[@class="panel-group accordion"]/div['+str(j)+']/div[2]/div/table/tbody/tr['+str(k)+']/td[2]/text()'
                path_role2 = '//*[@class="panel-group accordion"]/div['+str(j)+']/div[2]/div/table/tbody/tr['+str(k)+']/td[2]/span/text()'
                path_role3 = '//*[@class="panel-group accordion"]/div['+str(j)+']/div[2]/div/table/tbody/tr['+str(k)+']/td[2]/span/span/text()'
                path_role4 = '//*[@class="panel-group accordion"]/div['+str(j)+']/div[2]/div/table/tbody/tr['+str(k)+']/td[2]/p/text()'
                path_role5 = '//*[@class="panel-group accordion"]/div['+str(j)+']/div[2]/div/table/tbody/tr['+str(k)+']/td[2]/p/span/text()'
                path_role6 = '//*[@class="panel-group accordion"]/div['+str(j)+']/div[2]/div/table/tbody/tr['+str(k)+']/td[2]/a/text()'
                if tree.xpath(path_role) != [] and tree.xpath(path_role)[0] not in ['\n', '\xa0']:
                    role.append(tree.xpath(path_role)[0])
                elif tree.xpath(path_role2) != []:
                    role.append(tree.xpath(path_role2)[0])
                elif tree.xpath(path_role3) != []:
                    role.append(tree.xpath(path_role3)[0])
                elif tree.xpath(path_role4) != []:
                    role.append(tree.xpath(path_role4)[0])
                elif tree.xpath(path_role5) != []:
                    role.append(tree.xpath(path_role5)[0])
                elif tree.xpath(path_role6) != []:
                    role.append(tree.xpath(path_role6)[0])
                else: 
                    role.append("")
                college.append(college_name)
                department.append(departments[i])
    for j in [7]:
        for k in range(1,500):
            path_name = '//*[@class="panel-group accordion"]/div['+str(j)+']/div[2]/div/table/tbody/tr['+str(k)+']/td[1]/a/text()'
            if tree.xpath(path_name) != []:
                name.append(tree.xpath(path_name)[0])
                path_url = '//*[@class="panel-group accordion"]/div['+str(j)+']/div[2]/div/table/tbody/tr['+str(k)+']/td[1]/a/@href'
                try:
                    website.append(tree.xpath(path_url)[0])
                except:
                    website.append("") 
                role.append("PhD Student")
                college.append(college_name)
                department.append(departments[i])
                
ssis = pd.DataFrame(list(zip(college, department, name, role, website)),
                              columns=['college', 'department', 'name', 'role', 'url'])
ssis

ssis.department.value_counts()

# 2) Merge and save the data

staff = pd.concat([businessschool, cemps, humanities, lifesciences, medicine, ssis])
staff

staff.college.value_counts()

# Create names for matching
staff["namelast"] = staff.name.str.replace("Professor |Dr ", "", regex = True).str.split().apply(lambda x: x[-1])
staff["namefirst"] = staff.name.str.replace("Professor |Dr ", "", regex = True).str.split().apply(lambda x: x[0])
staff["identifier1"] = staff.namelast + ", " + staff.namefirst.str[0]
staff["identifier2"] = staff.namelast + ", " + staff.namefirst
# Rename url to profile
staff = staff.rename(columns={"url": "profile",})

# Drop all exact duplicates
staff = staff.drop_duplicates(subset=None, keep='first', inplace=False)
# Drop partial duplicates (name/department) where the role is "missing" in one of the observations
staff = staff.loc[~((staff.duplicated(['name', 'department'], keep=False)) & (staff.role.isnull()))]
# Drop all remaining department duplicates and keep last role
staff = staff.drop_duplicates(subset=['name', 'department'], keep='last', inplace=False)
# Drop partial duplicates (name/college) where the role is "missing" in last of the observations
staff = staff.loc[~((staff.duplicated(['name', 'college'], keep='last')) & (staff.role.isnull()))]
staff = staff.loc[~((staff.duplicated(['name', 'college'], keep='first')) & (staff.role.isnull()))]
# Drop partial duplicates (name/college) where one role is PhD and the othes is new role
staff = staff.loc[~((staff.duplicated(['name', 'college'], keep= False)) & (staff.role.str.match('PhD')))]
# Drop officers/ student recruitment managers / administrators / coordinators
staff = staff.loc[~staff.role.str.match('Web Officer', na=False)]
staff = staff.loc[~((staff.duplicated(['name', 'college'], keep= False)) & (staff.role.str.contains('Officer')))]
staff = staff.loc[~staff.role.str.contains('(?:Experimental|Scientific|Support|Political) Officer', na=False, regex = True)]
staff = staff.loc[~staff.role.str.contains('Student Recruitment', na=False, regex = True)]
staff = staff.loc[~staff.role.str.contains('Administrator|Coordinator|Marketing, Recruitment, Communications & Global Business Partner', na=False, regex = True)]
# Drop all remaining college duplicates and keep last role
staff = staff.drop_duplicates(subset=['name', 'college'], keep='first', inplace=False)

staff.to_csv('UoE_staff.csv', index = False)
# Last saved on 2022-04-01
