==> Code/02_UoE_ORE_publications.ipynb <==
import requests       # send requests to web server
# from bs4 import BeautifulSoup
from lxml import html # parse HTML
import json           # store data as json file
import re             # regular expressions
import os             # access directories
import pandas as pd   # dataframes
from tqdm import tqdm # create progress bar (for i in tqdm(list))
os.chdir('../Data')

## Scrape all URLS

%%time
# Get all article urls (2039 pages)
urls = []
for i in range(1,22):
    path = 'https://ore.exeter.ac.uk/repository/discover?rpp=10&etal=0&group_by=none&page='+str(i)+'&filtertype_0=type&filter_relational_operator_0=equals&filter_0=Article'
    page = requests.get(path)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(path)
    for url in tree.xpath('//a/@href'):
        if "https://ore.exeter.ac.uk/repository/handle/" in url:
            urls.append(url)
                                
articles = pd.DataFrame(urls, columns = ['url'])     
articles['type'] = 'article'  
print(len(urls), "urls were scraped.\n")

%%time
# Get all book urls (7 pages)
urls = []
for i in range(1,8):
    path = 'https://ore.exeter.ac.uk/repository/discover?rpp=10&etal=0&group_by=none&page='+str(i)+'&filtertype_0=type&filter_relational_operator_0=equals&filter_0=Book'
    page = requests.get(path)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(path)
    for url in tree.xpath('//a/@href'):
        if "https://ore.exeter.ac.uk/repository/handle/" in url:
            urls.append(url)
                                
books = pd.DataFrame(urls, columns = ['url'])     
books['type'] = 'book'
print(len(urls), "urls were scraped.\n")

%%time
# Get all book chapter urls (112 pages)
urls = []
for i in range(1,113):
    path = 'https://ore.exeter.ac.uk/repository/discover?rpp=10&etal=0&group_by=none&page='+str(i)+'&filtertype_0=type&filter_relational_operator_0=equals&filter_0=Book+chapter'
    page = requests.get(path)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(path)
    for url in tree.xpath('//a/@href'):
        if "https://ore.exeter.ac.uk/repository/handle/" in url:
            urls.append(url)
                                
bookchapters = pd.DataFrame(urls, columns = ['url'])     
bookchapters['type'] = 'book chapter'
print(len(urls), "urls were scraped.\n")

%%time
# Get all conference paper urls (87 pages)
urls = []
for i in range(1,88):
    path = 'https://ore.exeter.ac.uk/repository/discover?rpp=10&etal=0&group_by=none&page='+str(i)+'&filtertype_0=type&filter_relational_operator_0=equals&filter_0=Conference+paper'
    page = requests.get(path)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(path)
    for url in tree.xpath('//a/@href'):
        if "https://ore.exeter.ac.uk/repository/handle/" in url:
            urls.append(url)
                                
conferencepapers = pd.DataFrame(urls, columns = ['url'])     
conferencepapers['type'] = 'conferencepaper'
print(len(urls), "urls were scraped.\n")

%%time
# Get all thesis/dissertation urls (441 pages)
urls = []
for i in range(1,442):
    path = 'https://ore.exeter.ac.uk/repository/discover?rpp=10&etal=0&group_by=none&page='+str(i)+'&filtertype_0=type&filter_relational_operator_0=equals&filter_0=Thesis+or+dissertation'
    page = requests.get(path)
    tree = html.fromstring(page.content)
    tree.make_links_absolute(path)
    for url in tree.xpath('//a/@href'):
        if "https://ore.exeter.ac.uk/repository/handle/" in url:
            urls.append(url)
                                
theses = pd.DataFrame(urls, columns = ['url'])     
theses['type'] = 'thesis/dissertation'
print(len(urls), "urls were scraped.\n")

del(urls)

# Takes arouynd 8 minutes per 100 pages
# CPU times: user 7.56 s, sys: 263 ms, total: 7.83 s
# Wall time: 8min 1s

ore = pd.concat([articles, books, bookchapters, conferencepapers, theses])
ore.to_csv("ORE_urls.csv")

## Scrape metadata

ore

%%time
college = []
department = []
authors = []
date_accessioned = []
date_issued = []
abstract = []
title = []
DOIs = []

for url in ore.url:
    path = url + '?show=full'
    page = requests.get(path)
    tree = html.fromstring(page.content)
    college.append(tree.xpath('/html/body/div[1]/div/ul/li[2]/a/text()')[0])
    department.append(tree.xpath('/html/body/div[1]/div/ul/li[3]/a/text()')[0])
    author_list = []
    date_accessioned_switch = 0
    date_issued_switch = 0
    title_list_switch = 0
    abstract_list_switch = 0
    doi_list_switch = 0
    for i in range(0, len(tree.xpath('//*[@class="label-cell"]/text()'))):
        if tree.xpath('//*[@class="label-cell"]/text()')[i] == "dc.contributor.author":
            author_list.append(tree.xpath('//td[@class="word-break"]/text()')[i])
        if date_accessioned_switch == 0 and tree.xpath('//*[@class="label-cell"]/text()')[i] == "dc.date.accessioned":
            date_accessioned.append(tree.xpath('//td[@class="word-break"]/text()')[i])
            date_accessioned_switch = 1
        if date_issued_switch == 0 and tree.xpath('//*[@class="label-cell"]/text()')[i] == "dc.date.issued":
            date_issued.append(tree.xpath('//td[@class="word-break"]/text()')[i])
            date_issued_switch = 1
        if title_list_switch == 0 and tree.xpath('//*[@class="label-cell"]/text()')[i] == "dc.title":
            title.append(tree.xpath('//td[@class="word-break"]/text()')[i])
            title_list_switch = 1
        if abstract_list_switch == 0 and tree.xpath('//*[@class="label-cell"]/text()')[i] == "dc.description.abstract":
            abstract.append(tree.xpath('//td[@class="word-break"]/text()')[i])
            abstract_list_switch = 1
        if doi_list_switch == 0 and tree.xpath('//*[@class="label-cell"]/text()')[i] == "dc.identifier.doi":
            DOIs.append(tree.xpath('//td[@class="word-break"]/text()')[i])
            doi_list_switch = 1
    authors.append(author_list)
    if date_accessioned_switch == 0:
        date_accessioned.append("")
    if date_issued_switch == 0:
        date_issued.append("")
    if title_list_switch == 0:
        title.append("")
    if abstract_list_switch == 0:
        abstract.append("")
    if doi_list_switch == 0:
        DOIs.append("")

# 1.52 min for 66 urls
# CPU times: user 7.23 s, sys: 167 ms, total: 7.4 s
# Wall time: 1min 52s
# >> Might take 8 hours for the full 22935 urls

df = pd.concat([ore.reset_index(drop=True), 
                pd.DataFrame(list(zip(college, department, authors, date_accessioned, date_issued, title, abstract, doi)),
                             columns=['college', 'department', 'authors', 'date_accessioned', 'date_issued', 'title', 'abstract', 'doi'])] , axis = 1)
df

df.to_csv("ORE_data.csv", index = False)
# Last saved on 2022-04-05 (with DOIs)
