==> Code/03_UoE_match researchers with ORE.ipynb <==
import re             # regular expressions
import os             # access directories
import pandas as pd   # dataframes
from tqdm import tqdm # create progress bar (for i in tqdm(list))
os.chdir('../Data')

## Load data

staff = pd.read_csv('UoE_staff.csv')
staff

# Prepare data for matching
staff.identifier1 = staff.identifier1.str.lower()
staff.identifier2 = staff.identifier2.str.lower()
staff.college[staff.college == 'The Business School'] = 'Business School'

ore = pd.read_csv('ORE_data.csv')
ore['identifiers'] = ore.authors.map(lambda x: x.lower())
ore

# Check proportion of missing information in the ORE data
(ore.isnull().sum()*100/len(ore)).round(1)

# Check which colleges in the ORE data do not match the staff data
for col in ore.college.unique():
    if not col in staff.college.unique():
     print(col)

# Check which department in the ORE data do not match the staff data
for dep in ore.department.unique():
    if not dep in staff.department.unique():
     print(dep)

for dep in staff.department.unique():
    if not dep in ore.department.unique():
     print(dep)

## Match staff members to metadata

# %%time
# # Match by department
# authors = []
# dates = []
# titles = []
# abstracts = []
# urls = []
# fields = []
# publications = []

# for i, name in enumerate(staff.name):
#     authors_temp = []
#     dates_temp = []
#     titles_temp = []
#     abstracts_temp = []
#     url_temp = []
#     field_temp = []
#     count = 0
#     for j, title in enumerate(ore.title):
#         if ((staff.identifier1[i] in ore.identifiers[j]) or (staff.identifier2[i] in ore.identifiers[j])) and staff.college[i] in ore.college[j]:
#             authors_temp.append(ore.authors[j])
#             dates_temp.append(ore.date_issued[j])
#             titles_temp.append(ore.title[j])
#             abstracts_temp.append(ore.abstract[j])
#             url_temp.append(ore.url[j])
#             field_temp.append(ore.department[j])
#             count +=1
#     publications.append(count)
#     authors.append(authors_temp)
#     dates.append(dates_temp)
#     titles.append(titles_temp)
#     abstracts.append(abstracts_temp)
#     urls.append(url_temp) 
#     fields.append(field_temp)

%%time
# Match by name and college
authors = []
dates = []
titles = []
abstracts = []
urls = []
fields = []
types = []
publications = []

for i, name in enumerate(staff.name):
    authors_temp = []
    dates_temp = []
    titles_temp = []
    abstracts_temp = []
    url_temp = []
    field_temp = []
    type_temp = []
    count = 0
    for j, title in enumerate(ore.title):
        if ((staff.identifier1[i] in ore.identifiers[j]) or (staff.identifier2[i] in ore.identifiers[j])) and (staff.college[i] in ore.college[j] or ore.college[j] == "Doctoral College"):
            authors_temp.append(ore.authors[j])
            dates_temp.append(ore.date_issued[j])
            titles_temp.append(ore.title[j])
            abstracts_temp.append(ore.abstract[j])
            url_temp.append(ore.url[j])
            field_temp.append(ore.department[j])
            type_temp.append(ore.type[j])
            count +=1
    publications.append(count)
    authors.append(authors_temp)
    dates.append(dates_temp)
    titles.append(titles_temp)
    abstracts.append(abstracts_temp)
    urls.append(url_temp) 
    fields.append(field_temp)
    types.append(type_temp)

df = pd.concat([staff.reset_index(drop=True), 
                pd.DataFrame(list(zip(publications, authors, dates, titles, abstracts, urls, fields, types)),
                             columns=['publications', 'authors', 'dates', 'titles', 'abstracts', 'urls', 'fields', 'types'])] , axis = 1)
df

# Manually fix wrong matches for authors with 40+ publications

# # Ali A
# df.loc[1243,'authors':'types'] = df.loc[1243,'authors':'types'].apply(lambda x: list())
# df.loc[1243,'publications'] = len(df.loc[1243,'authors'])

# # Rod and Robert Taylor
# taylor_robert = []
# taylor_rod = []
# for i, authors in enumerate(df.loc[3245]['authors']):
#     if len(re.findall("Taylor, Robert|Taylor, R'|Taylor, RW", authors)) != 0:
#         taylor_robert.append(i)
#     if len(re.findall("Taylor, Rod S|Taylor, RS", authors)) != 0:
#         taylor_rod.append(i)
# df.loc[3245,'authors':'types'] = df.loc[3245,'authors':'types'].apply(lambda x: [x[i] for i in taylor_rod])
# df.loc[3245,'publications'] = len(df.loc[3245,'authors'])
# df.loc[3246,'authors':'types'] = df.loc[3246,'authors':'types'].apply(lambda x: [x[i] for i in taylor_robert])
# df.loc[3246,'publications'] = len(df.loc[3246,'authors'])

# Yang, Yaan, and Yinchen Liu
# liu_yang = []
# liu_yaan = []
# for i, authors in enumerate(df.loc[443]['authors']):
#     # All of Yaan Liu's articles are coauthored with Ghita
#     if len(re.findall("Ghita", authors)) != 0:
#         liu_yaan.append(i)
#     else:
#         liu_yang.append(i)
# df.loc[443,'authors':'types'] = df.loc[443,'authors':'types'].apply(lambda x: [x[i] for i in liu_yang])
# df.loc[443,'publications'] = len(df.loc[443,'authors'])
# df.loc[737,'authors':'types'] = df.loc[737,'authors':'types'].apply(lambda x: [x[i] for i in liu_yaan])
# df.loc[737,'publications'] = len(df.loc[737,'authors'])
# df.loc[1525,'authors':'types'] = df.loc[1525,'authors':'types'].apply(lambda x: list())
# df.loc[1525,'publications'] = len(df.loc[1525,'authors'])

# # Craig and C. David Wright
# wright_c_david = []
# for i, authors in enumerate(df.loc[477]['authors']):
#     if len(re.findall("Wright, C. David|Wright, CD|Wright, C'|David Wright, C", authors)) != 0:
#         wright_c_david.append(i)
# df.loc[477,'authors':'types'] = df.loc[477,'authors':'types'].apply(lambda x: [x[i] for i in wright_c_david])
# df.loc[477,'publications'] = len(df.loc[477,'authors'])
# df.loc[1065,'authors':'types'] = df.loc[1065,'authors':'types'].apply(lambda x: list())
# df.loc[1065,'publications'] = len(df.loc[1065,'authors'])

# # Hani Y
# df.loc[2827,'authors':'types'] = df.loc[2827,'authors':'types'].apply(lambda x: list())
# df.loc[2827,'publications'] = len(df.loc[2827,'authors'])

# # Xinqi and Xiang Li
# df.loc[675,'authors':'types'] = df.loc[675,'authors':'types'].apply(lambda x: list())
# df.loc[675,'publications'] = len(df.loc[675,'authors'])
# df.loc[1524,'authors':'types'] = df.loc[1524,'authors':'types'].apply(lambda x: list())
# df.loc[1524,'publications'] = len(df.loc[1524,'authors'])

# # Yuxiao, Yi, and Ying Wang
# wang_ying = []
# for i, authors in enumerate(df.loc[762]['authors']):
#     # All of Ying Wang's articles are coauthored with Dai, K
#     if len(re.findall("Dai, K", authors)) != 0:
#         wang_ying.append(i)     
# df.loc[762,'authors':'types'] = df.loc[762,'authors':'types'].apply(lambda x: [x[i] for i in wang_ying])
# df.loc[762,'publications'] = len(df.loc[762,'authors'])
# df.loc[1224,'authors':'types'] = df.loc[1224,'authors':'types'].apply(lambda x: list())
# df.loc[1224,'publications'] = len(df.loc[1224,'authors'])
# df.loc[1353,'authors':'types'] = df.loc[1353,'authors':'types'].apply(lambda x: list())
# df.loc[1353,'publications'] = len(df.loc[1353,'authors'])

# Yujia Zhu
# df.loc[1080,'authors':'types'] = df.loc[1080,'authors':'types'].apply(lambda x: list())
# df.loc[1080,'publications'] = len(df.loc[1080,'authors'])

# # Sarah Bell and Sian de Bell
# bell_sarah = []
# debell_sian = []
# for i, authors in enumerate(df.loc[2855]['authors']):
#     if len(re.findall("Bell, SL|Bell, Sarah L|^Bell, S", authors)) != 0:
#         bell_sarah.append(i)
#     elif len(re.findall("de Bell", authors)) != 0:
#         debell_sian.append(i)
# df.loc[2855,'authors':'types'] = df.loc[2855,'authors':'types'].apply(lambda x: [x[i] for i in bell_sarah])
# df.loc[2855,'publications'] = len(df.loc[2855,'authors'])
# df.loc[2919,'authors':'types'] = df.loc[2919,'authors':'types'].apply(lambda x: [x[i] for i in debell_sian])
# df.loc[2919,'publications'] = len(df.loc[2919,'authors'])

# # Rebecca and Richard Smith
# smith_rebecca = []
# smith_richard = []
# for i, authors in enumerate(df.loc[3224]['authors']):
#     if len(re.findall("Smith, RG|Smith, Rebecca", authors)) != 0:
#         smith_rebecca.append(i)
#     elif len(re.findall("Smith, R'|Smith, RD", authors)) != 0:
#         smith_richard.append(i)
# df.loc[3224,'authors':'types'] = df.loc[3224,'authors':'types'].apply(lambda x: [x[i] for i in smith_rebecca])
# df.loc[3224,'publications'] = len(df.loc[3224,'authors'])
# df.loc[3225,'authors':'types'] = df.loc[3225,'authors':'types'].apply(lambda x: [x[i] for i in smith_richard])
# df.loc[3225,'publications'] = len(df.loc[3225,'authors'])

# Manually fix wrong matches for authors with 30+ publications

# Abigail and Anne-Marie Russell
russell_abigail = []
russell_anne = []
for i, authors in enumerate(df.loc[3201]['authors']):
    if len(re.findall("Russell, AM|Russell, A-M", authors)) != 0:
        russell_anne.append(i)
    else:
        russell_abigail.append(i)
df.loc[3201,'authors':'types'] = df.loc[3201,'authors':'types'].apply(lambda x: [x[i] for i in russell_abigail])
df.loc[3201,'publications'] = len(df.loc[3201,'authors'])
df.loc[3202,'authors':'types'] = df.loc[3202,'authors':'types'].apply(lambda x: [x[i] for i in russell_anne])
df.loc[3202,'publications'] = len(df.loc[3202,'authors'])

# Ali A
df.loc[1243,'authors':'types'] = df.loc[1243,'authors':'types'].apply(lambda x: list())
df.loc[1243,'publications'] = len(df.loc[1243,'authors'])

# Aimee and Anna Murray
murray_aimee = []
murray_anna = []
for i, authors in enumerate(df.loc[3125]['authors']):
    if len(re.findall("Murray, AK|Bendall, R", authors)) != 0:
        murray_aimee.append(i)
    else:
        murray_anna.append(i)
df.loc[3125,'authors':'types'] = df.loc[3125,'authors':'types'].apply(lambda x: [x[i] for i in murray_aimee])
df.loc[3125,'publications'] = len(df.loc[3125,'authors'])
df.loc[3126,'authors':'types'] = df.loc[3126,'authors':'types'].apply(lambda x: [x[i] for i in murray_anna])
df.loc[3126,'publications'] = len(df.loc[3126,'authors'])

# Chao and Chi Zhang
df.loc[1074,'authors':'types'] = df.loc[1074,'authors':'types'].apply(lambda x: list())
df.loc[1074,'publications'] = len(df.loc[1074,'authors'])
df.loc[1239,'authors':'types'] = df.loc[1239,'authors':'types'].apply(lambda x: list())
df.loc[1239,'publications'] = len(df.loc[1239,'authors'])

# Craig and C. David Wright
wright_c_david = []
for i, authors in enumerate(df.loc[477]['authors']):
    if len(re.findall("Wright, C. David|Wright, CD|Wright, C'|David Wright, C", authors)) != 0:
        wright_c_david.append(i)
df.loc[477,'authors':'types'] = df.loc[477,'authors':'types'].apply(lambda x: [x[i] for i in wright_c_david])
df.loc[477,'publications'] = len(df.loc[477,'authors'])
df.loc[1065,'authors':'types'] = df.loc[1065,'authors':'types'].apply(lambda x: list())
df.loc[1065,'publications'] = len(df.loc[1065,'authors'])

# Daniel and Derek Partridge
partridge_daniel = []
partridge_derek = []
for i, authors in enumerate(df.loc[540]['authors']):
    if len(re.findall("Partridge, Derek", authors)) != 0:
        partridge_derek.append(i)
    else:
        partridge_daniel.append(i)
df.loc[540,'authors':'types'] = df.loc[540,'authors':'types'].apply(lambda x: [x[i] for i in partridge_daniel])
df.loc[540,'publications'] = len(df.loc[540,'authors'])
df.loc[1569,'authors':'types'] = df.loc[1569,'authors':'types'].apply(lambda x: [x[i] for i in partridge_derek])
df.loc[1569,'publications'] = len(df.loc[1569,'authors'])

# Hani Y
df.loc[2827,'authors':'types'] = df.loc[2827,'authors':'types'].apply(lambda x: list())
df.loc[2827,'publications'] = len(df.loc[2827,'authors'])

# Junhao, Jingyi and Jin Wang
wang_jin = []
for i, titles in enumerate(df.loc[1059]['titles']):
    if len(re.findall("Data Augmentation|Learning", titles)) != 0:
        wang_jin.append(i)
df.loc[1059,'authors':'types'] = df.loc[1059,'authors':'types'].apply(lambda x: [x[i] for i in wang_jin])
df.loc[1059,'publications'] = len(df.loc[1059,'authors'])
df.loc[1108,'authors':'types'] = df.loc[1108,'authors':'types'].apply(lambda x: list())
df.loc[1108,'publications'] = len(df.loc[1108,'authors'])
df.loc[1225,'authors':'types'] = df.loc[1225,'authors':'types'].apply(lambda x: list())
df.loc[1225,'publications'] = len(df.loc[1225,'authors'])

# Katrina and Kerry Brown
brown_katrina = []
brown_kerry = []
junk = []
for i, authors in enumerate(df.loc[2485]['authors']):
    if len(re.findall("Adger|Baggio|Chaigneau|Galafassi|Huke|Morrison|Quinn|Schoon", authors)) != 0 or authors == "['Brown, K']":
        brown_katrina.append(i)
    elif len(re.findall("Venkateshmurthy", authors)) != 0:
        brown_kerry.append(i)
df.loc[2485,'authors':'types'] = df.loc[2485,'authors':'types'].apply(lambda x: [x[i] for i in brown_katrina])
df.loc[2485,'publications'] = len(df.loc[2485,'authors'])
df.loc[2698,'authors':'types'] = df.loc[2698,'authors':'types'].apply(lambda x: [x[i] for i in brown_kerry])
df.loc[2698,'publications'] = len(df.loc[2698,'authors'])

# Rebecca and Richard Smith
smith_rebecca = []
smith_richard = []
for i, authors in enumerate(df.loc[3224]['authors']):
    if len(re.findall("Smith, RG|Smith, Rebecca", authors)) != 0:
        smith_rebecca.append(i)
    elif len(re.findall("Smith, R'|Smith, RD", authors)) != 0:
        smith_richard.append(i)
df.loc[3224,'authors':'types'] = df.loc[3224,'authors':'types'].apply(lambda x: [x[i] for i in smith_rebecca])
df.loc[3224,'publications'] = len(df.loc[3224,'authors'])
df.loc[3225,'authors':'types'] = df.loc[3225,'authors':'types'].apply(lambda x: [x[i] for i in smith_richard])
df.loc[3225,'publications'] = len(df.loc[3225,'authors'])

# Rod and Robert Taylor
taylor_robert = []
taylor_rod = []
for i, authors in enumerate(df.loc[3245]['authors']):
    if len(re.findall("Taylor, Robert|Taylor, R'|Taylor, RW", authors)) != 0:
        taylor_robert.append(i)
    if len(re.findall("Taylor, Rod S|Taylor, RS", authors)) != 0:
        taylor_rod.append(i)
df.loc[3245,'authors':'types'] = df.loc[3245,'authors':'types'].apply(lambda x: [x[i] for i in taylor_rod])
df.loc[3245,'publications'] = len(df.loc[3245,'authors'])
df.loc[3246,'authors':'types'] = df.loc[3246,'authors':'types'].apply(lambda x: [x[i] for i in taylor_robert])
df.loc[3246,'publications'] = len(df.loc[3246,'authors'])

# Sarah Bell and Sian de Bell
bell_sarah = []
debell_sian = []
for i, authors in enumerate(df.loc[2855]['authors']):
    if len(re.findall("Bell, SL|Bell, Sarah L|^Bell, S", authors)) != 0:
        bell_sarah.append(i)
    elif len(re.findall("de Bell", authors)) != 0:
        debell_sian.append(i)
df.loc[2855,'authors':'types'] = df.loc[2855,'authors':'types'].apply(lambda x: [x[i] for i in bell_sarah])
df.loc[2855,'publications'] = len(df.loc[2855,'authors'])
df.loc[2919,'authors':'types'] = df.loc[2919,'authors':'types'].apply(lambda x: [x[i] for i in debell_sian])
df.loc[2919,'publications'] = len(df.loc[2919,'authors'])

# Sigong Zhang
df.loc[767,'authors':'types'] = df.loc[767,'authors':'types'].apply(lambda x: list())
df.loc[767,'publications'] = len(df.loc[767,'authors'])

# Yang, Yaan, and Yinchen Liu
liu_yang = []
liu_yaan = []
for i, authors in enumerate(df.loc[443]['authors']):
    # All of Yaan Liu's articles are coauthored with Ghita
    if len(re.findall("Ghita", authors)) != 0:
        liu_yaan.append(i)
    else:
        liu_yang.append(i)
df.loc[443,'authors':'types'] = df.loc[443,'authors':'types'].apply(lambda x: [x[i] for i in liu_yang])
df.loc[443,'publications'] = len(df.loc[443,'authors'])
df.loc[737,'authors':'types'] = df.loc[737,'authors':'types'].apply(lambda x: [x[i] for i in liu_yaan])
df.loc[737,'publications'] = len(df.loc[737,'authors'])
df.loc[1525,'authors':'types'] = df.loc[1525,'authors':'types'].apply(lambda x: list())
df.loc[1525,'publications'] = len(df.loc[1525,'authors'])

# Yujia Zhu
df.loc[1080,'authors':'types'] = df.loc[1080,'authors':'types'].apply(lambda x: list())
df.loc[1080,'publications'] = len(df.loc[1080,'authors'])

# Yuxiao, Yi, and Ying Wang
wang_ying = []
for i, authors in enumerate(df.loc[762]['authors']):
    # All of Ying Wang's articles are coauthored with Dai, K
    if len(re.findall("Dai, K", authors)) != 0:
        wang_ying.append(i)     
df.loc[762,'authors':'types'] = df.loc[762,'authors':'types'].apply(lambda x: [x[i] for i in wang_ying])
df.loc[762,'publications'] = len(df.loc[762,'authors'])
df.loc[1224,'authors':'types'] = df.loc[1224,'authors':'types'].apply(lambda x: list())
df.loc[1224,'publications'] = len(df.loc[1224,'authors'])
df.loc[1353,'authors':'types'] = df.loc[1353,'authors':'types'].apply(lambda x: list())
df.loc[1353,'publications'] = len(df.loc[1353,'authors'])

# Xinqi and Xiang Li
df.loc[675,'authors':'types'] = df.loc[675,'authors':'types'].apply(lambda x: list())
df.loc[675,'publications'] = len(df.loc[675,'authors'])
df.loc[1524,'authors':'types'] = df.loc[1524,'authors':'types'].apply(lambda x: list())
df.loc[1524,'publications'] = len(df.loc[1524,'authors'])

# A lot of duplicates remain, but these are too many to fix manually
df[(df.duplicated(subset=['namelast', 'publications'], keep = False)) & (df.publications>0) ].sort_values(by=['publications', 'namelast'])

# Explode the data
df["id"] = df.index
df_long = df.set_index(['id']).apply(pd.Series.explode).reset_index()
df_long.authors = df_long.authors.str.strip('[]')
df_long

# Save/load data
df_long.to_csv('UoE_staff_publications_theses.csv', index = False)
