==> Code/05_Match_Methods_to_Abstracts.ipynb <==
import re             # regular expressions
import string         # string operations
import os             # access directories
import pandas as pd   # dataframes
from tqdm import tqdm # create progress bar (for i in tqdm(list))
tqdm.pandas()         # use progress_apply() instead of apply() for progress bar on pandas operations
import spacy          # Natural Language Processing library
from spaczz.matcher import FuzzyMatcher # fuzzy keyword matching

os.chdir('../Data')

# Load and prepare data

df = pd.read_csv('UoE_staff_publications_theses.csv')
df['abstracts_matching'] = ' ' + df.abstracts.str.replace('[{}]'.format(string.punctuation), ' ', regex = True).str.lower() + ' '
df.head()

methods = pd.read_csv('methods.csv')
methods['method_matching'] = ' ' + methods.method_lower.str.replace('[{}]'.format(string.punctuation), ' ', regex = True) + ' '
m = methods[['method', 'method_lower', 'method_matching']].drop_duplicates('method_matching').reset_index(drop = True)
m.head()

# spaczz: FuzzyMatcher

# Set up matcher
nlp = spacy.blank("en")
matcher = FuzzyMatcher(nlp.vocab, min_r2=90)
for i, method in enumerate(m.method):
    matcher.add(method, [nlp(m['method_lower'][i])])

# Preprocess the abstracts (This takes ~ 1hour)
df['abstract_nlp'] = df.abstracts.progress_apply(lambda x: nlp(x) if pd.notnull(x) else x)

# Match the methods (This takes ~ 20hours!)
df['matches'] = df.abstract_nlp.progress_apply(lambda x: matcher(x) if pd.notnull(x) else x)

# Extract the matched methods from the spaczz matches objects
df['methods'] = df.matches.progress_apply(lambda x: [i[0] for i in x] if type(x) == list else x)
df['methods_tokens'] = df.progress_apply(lambda x: [x['abstract_nlp'][j[0]:j[1]] for j in [i[1:3] for i in x['matches']]] if type(x['matches']) == list else x['matches'], axis = 1)
df['match_ratios'] = df.matches.progress_apply(lambda x: [i[3] for i in x] if type(x) == list else x)

# Save the data
df.to_csv('UoE_staff_publications_theses_spaczz_matches.csv')
df.to_json("UoE_staff_publications_theses_spaczz_matches.json", default_handler=str) # Saving as json handles the nested lists better

# fuzzysearch: find_near_matches

# %time
# index_list = []
# match_list = []
# my_string = methods.method_matching[0]
# for i, abstract in enumerate(tqdm(df.abstracts_matching)):
#     try:
#         match = find_near_matches(my_string, abstract, max_l_dist=1)
#         if match != []:
#             index_list.append(i)
#             match_list.append(match)
#     except:
#         pass
# #         match_list.append(None)

# index_list

# match_list

# %time
# df['match'] = df[df.abstracts_matching.isnull() == False].abstracts_matching.apply(lambda x: find_near_matches(my_string, x, max_l_dist=1))
